testing/web-platform/mozilla/tests/binast/domain-restrict-included.https.html

<!DOCTYPE html>
<title>Check we can load BinAST if the host is included in the list</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    setup({allow_uncaught_exception: true});

    var hadScriptLoadError = false;
    function setLoadError() {
      // Load error happens if the server side throws an exception,
      // for 'expect_accept' check on server side.
      hadScriptLoadError = true;
    }

    var hadOtherError = false;
    function setOtherError() {
      hadOtherError = true;
    }

    window.addEventListener("error", setOtherError);

    const test_load = async_test("Check we can load BinAST if the host is included in the list");
    window.addEventListener("load", test_load.step_func_done(ev => {
      assert_equals(hadScriptLoadError, false, "Didn't expect a load error event");
      assert_equals(hadOtherError, false, "Didn't expect other error event");
      assert_equals(binASTLoaded, true, "Expected to load BinAST version");
    }));

</script>
<script src="./serve.py?name=small&amp;expect_accept=true" onerror="setLoadError()"></script>
